pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
django-phonenumber-field = "^8.1.0"
djangorestframework = "^3.16.0"
phonenumbers = "^9.0.10"
pillow = "^11.3.0"
dotenv = "^0.9.9"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
